# besserlesenschreiben/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/request-code": {
      "post": {
        "operationId": "AuthController_requestCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/verify": {
      "post": {
        "operationId": "AuthController_verify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                  }
                },
                "required": [
                  "email",
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "isNewAccount": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "token",
                    "isNewAccount"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "ProfilesController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "account": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "email"
                      ],
                      "additionalProperties": false
                    },
                    "profiles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "buddy": {
                            "type": "string"
                          },
                          "goalPerWeek": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "soundOn": {
                            "type": "boolean"
                          },
                          "dyslexicFont": {
                            "type": "boolean"
                          },
                          "fontScale": {
                            "type": "number"
                          },
                          "stars": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "streakDays": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "unlockedUnit": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "buddy",
                          "goalPerWeek",
                          "soundOn",
                          "dyslexicFont",
                          "fontScale",
                          "stars",
                          "streakDays",
                          "unlockedUnit",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "account",
                    "profiles"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles": {
      "post": {
        "operationId": "ProfilesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 40
                  },
                  "buddy": {
                    "type": "string",
                    "enum": [
                      "nepo",
                      "stella"
                    ]
                  },
                  "goal": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 14
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "buddy": {
                          "type": "string"
                        },
                        "goalPerWeek": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "soundOn": {
                          "type": "boolean"
                        },
                        "dyslexicFont": {
                          "type": "boolean"
                        },
                        "fontScale": {
                          "type": "number"
                        },
                        "stars": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "streakDays": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "unlockedUnit": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "createdAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "buddy",
                        "goalPerWeek",
                        "soundOn",
                        "dyslexicFont",
                        "fontScale",
                        "stars",
                        "streakDays",
                        "unlockedUnit",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "profile"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/{id}": {
      "get": {
        "operationId": "ProfilesController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "buddy": {
                          "type": "string"
                        },
                        "goalPerWeek": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "soundOn": {
                          "type": "boolean"
                        },
                        "dyslexicFont": {
                          "type": "boolean"
                        },
                        "fontScale": {
                          "type": "number"
                        },
                        "stars": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "streakDays": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "unlockedUnit": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "createdAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "buddy",
                        "goalPerWeek",
                        "soundOn",
                        "dyslexicFont",
                        "fontScale",
                        "stars",
                        "streakDays",
                        "unlockedUnit",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    },
                    "settings": {
                      "type": "object",
                      "properties": {
                        "soundOn": {
                          "type": "boolean"
                        },
                        "dyslexicFont": {
                          "type": "boolean"
                        },
                        "fontScale": {
                          "type": "number"
                        },
                        "goalPerWeek": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "buddy": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "soundOn",
                        "dyslexicFont",
                        "fontScale",
                        "goalPerWeek",
                        "buddy"
                      ],
                      "additionalProperties": false
                    },
                    "stars": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "streak": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "profile",
                    "settings",
                    "stars",
                    "streak"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/{id}/settings": {
      "patch": {
        "operationId": "ProfilesController_updateSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "soundOn": {
                    "type": "boolean"
                  },
                  "dyslexicFont": {
                    "type": "boolean"
                  },
                  "fontScale": {
                    "type": "number",
                    "minimum": 0.8,
                    "maximum": 2
                  },
                  "goal": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 14
                  },
                  "buddy": {
                    "type": "string",
                    "enum": [
                      "nepo",
                      "stella"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "buddy": {
                          "type": "string"
                        },
                        "goalPerWeek": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "soundOn": {
                          "type": "boolean"
                        },
                        "dyslexicFont": {
                          "type": "boolean"
                        },
                        "fontScale": {
                          "type": "number"
                        },
                        "stars": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "streakDays": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "unlockedUnit": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "createdAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "buddy",
                        "goalPerWeek",
                        "soundOn",
                        "dyslexicFont",
                        "fontScale",
                        "stars",
                        "streakDays",
                        "unlockedUnit",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "profile"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "profiles"
        ]
      }
    },
...
  },
  "info": {
    "title": "besserlesenschreiben API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}